# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangent Blog OpenApi Demo Documentation",
        "description": "Tangent Blog Swagger OpenApi description",
        "contact": {
            "email": "sipho411@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://tangent_blog.test",
            "description": "Demo Tangent Blog API Server"
        }
    ],
    "paths": {
        "/createUser": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create a user",
                "description": "Creates a user",
                "operationId": "createUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/loginUser": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Log in an existing user",
                "description": "Logs in an existing user",
                "operationId": "loginUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get all categories.",
                "description": "Get all categories.",
                "operationId": "categoryList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/category/store": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Add a category.",
                "description": "Add a category.",
                "operationId": "storeCategory",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/category/{id}/show": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get category.",
                "description": "Get category.",
                "operationId": "categoryDetail",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/category/{id}/update": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Update a category.",
                "description": "Update a category.",
                "operationId": "categoryUpdate",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/category/{id}/destroy": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category.",
                "description": "Delete a category.",
                "operationId": "categoryDestroy",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/category/{keyword}/search": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Search categories.",
                "description": "Search categories.",
                "operationId": "categorySearch",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
}